requirements.txt
django==3.2.6
djangorestframework==3.12.4
django-cors-headers==3.7.0
django-filter==2.4.0
flake8==3.9.2
flake8-import-order==0.18.1
psycopg2-binary==2.9.1
daphne==3.0.2
sentry-sdk==1.3.1
pillow==8.4.0
django-solo==2.0.0
django-compressor==3.0
django-libsass==0.9
gunicorn==20.1.0
django-ckeditor==6.2.0